changed users endpoint to be tokens and moved login route to auth
