+++ sesame/webapp/api/auth.py
@@ -1,13 +1,19 @@
 import os
 from typing import Optional
 
+from argon2 import PasswordHasher
+from argon2.exceptions import VerifyMismatchError
 from common.database import DatabaseSessionFactory
-from common.models import Token, User
+from common.models import Token, User, UserLoginModel
 from dotenv import load_dotenv
 from fastapi import APIRouter, Depends, HTTPException, status
 from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
+from sqlalchemy import select, text
+from sqlalchemy.ext.asyncio import AsyncSession
 
 load_dotenv(override=True)
+
+ph = PasswordHasher()
 
 AuthProvider = None
 auth_provider = os.getenv("SESAME_AUTH_PROVIDER", None)
@@ -27,6 +33,76 @@ router = APIRouter(
 bearer_scheme = HTTPBearer(auto_error=False)
 
 default_session_factory = DatabaseSessionFactory()
+
+
+async def _authenticate_user(credentials: UserLoginModel, db_session: AsyncSession):
+    result = await db_session.execute(
+        text("SELECT * FROM check_rate_limit(:email, :max_attempts, :window_minutes)"),
+        {
+            "email": credentials.email,
+            "max_attempts": int(os.getenv("SESAME_MAX_LOGIN_ATTEMPTS", 5)),
+            "window_minutes": int(os.getenv("SESAME_RATE_LIMIT_WINDOW_MINUTES", 15)),
+        },
+    )
+    rate_limit_ok = result.scalar_one()
+    await db_session.commit()
+
+    if not rate_limit_ok:
+        raise Exception("Rate limit exceeded. Please try again later.")
+
+    result = await db_session.execute(
+        text("SELECT * FROM get_user_for_login(:email)"), {"email": credentials.email}
+    )
+    user = result.fetchone()
+
+    if not user:
+        raise Exception("Invalid email or password")
+
+    try:
+        ph.verify(user.password_hash, credentials.password)
+    except VerifyMismatchError:
+        raise Exception("Invalid email or password")
+
+    await db_session.execute(
+        text("SELECT set_current_user_id(:user_id)"), {"user_id": user.user_id}
+    )
+
+    await db_session.execute(
+        text("DELETE FROM login_attempts WHERE email = :email"),
+        {"email": credentials.email},
+    )
+
+    await db_session.commit()
+
+    return user
+
+
+@router.post("/login", name="Login with credentials")
+async def login_with_credentials(
+    credentials: UserLoginModel,
+):
+    async with default_session_factory() as db_session:
+        try:
+            user = await _authenticate_user(credentials, db_session)
+        except Exception as e:
+            raise HTTPException(status_code=400, detail=str(e))
+
+        # Get or create a token for the user
+        tokens = await db_session.execute(
+            select(Token).where(Token.user_id == user.user_id, Token.revoked.is_(False)).limit(1)
+        )
+        token = tokens.scalar_one_or_none()
+
+        if not token:
+            token = await Token.create_token_for_user(
+                user.user_id,
+                db_session,
+                title="Default Token",
+                expiration_minutes=int(os.getenv("SESAME_TOKEN_EXPIRY", 525600)),
+            )
+            await db_session.commit()
+
+        return {"success": True, "token": token.token}
 
 
 @router.get("/token")
