+++ sesame/webapp/dashboard/index.html
@@ -86,7 +86,7 @@
         this.isFetching = true;
         this.responseMessage = "";
         try {
-          const response = await fetch("/api/users/login", {
+          const response = await fetch("/api/auth/login", {
             method: "POST",
             headers: {
               "Content-Type": "application/json",
@@ -156,8 +156,8 @@
       async fetchTokens() {
         this.isFetchingTokens = true;
         try {
-          const response = await fetch("/api/users/tokens", {
-            method: "POST",
+          const response = await fetch("/api/tokens", {
+            method: "GET",
             headers: {
               "Content-Type": "application/json",
               Authorization: `Bearer ${this.token}`,
@@ -186,7 +186,7 @@
         this.isCreateTokenModalActive = false;
         this.isFetchingTokens = true;
         try {
-          const response = await fetch("/api/users/token", {
+          const response = await fetch("/api/tokens/token", {
             method: "POST",
             headers: {
               "Content-Type": "application/json",
@@ -223,7 +223,7 @@
         }
 
         try {
-          const response = await fetch("/api/users/revoke_token", {
+          const response = await fetch("/api/tokens/revoke", {
             method: "POST",
             headers: {
               "Content-Type": "application/json",
